+++ examples/drop_down/one_option/index.jsx
@@ -1,0 +1,38 @@
+import { useRef, useState } from 'react';
+import { DropDown } from '../../../src';
+
+const options = ['There can be only one'];
+
+export function Example() {
+  const [value, setValue] = useState(null);
+  const ref = useRef();
+  return (
+    <>
+      <div
+        className="label"
+        onClick={() => ref.current.focus()}
+        id="drop-down-label"
+      >
+        Drop down
+      </div>
+      <DropDown
+        ref={ref}
+        id="drop-down"
+        aria-labelledby="drop-down-label"
+        value={value}
+        onValue={setValue}
+        options={options}
+      >
+        {value}
+      </DropDown>
+
+      <label htmlFor="output">Current value</label>
+      <output
+        htmlFor="drop-down"
+        id="output"
+      >
+        {JSON.stringify(value, undefined, ' ')}
+      </output>
+    </>
+  );
+}